BUG: testing on windows
- we are passing builds which actually have an error
- fix the small dtype issues

Author: Jeff Reback <jeff@reback.net>

Closes #15445 from jreback/windows and squashes the following commits:

a5b7fb3 [Jeff Reback] change integer to power comparisions
eab15c4 [Jeff Reback] don't force remove pandas
cf3b9bd [Jeff Reback] more windows fixing
efe6a76 [Jeff Reback] add cython to build
8194e63 [Jeff Reback] don't use appveyor recipe, just build inplace
e064825 [Jeff Reback] TST: resample dtype issue xref #15418
10d9b26 [Jeff Reback] TST: run windows tests so failures show up in appeveyor

+++ pandas/tests/tseries/test_resample.py
@@ -1944,7 +1944,7 @@ class TestDatetimeIndex(Base, tm.TestCase):
         index = pd.date_range('1-1-2000', '2-15-2000', freq='h')
         index2 = pd.date_range('4-15-2000', '5-15-2000', freq='h')
         index3 = index.append(index2)
-        s = pd.Series(range(len(index3)), index=index3)
+        s = pd.Series(range(len(index3)), index=index3, dtype='int64')
         r = s.resample('M')
 
         # Since all elements are unique, these should all be the same
